Refactor calls to GDrive 'list' API calls
In order to avoid a potential timeout and lose all progress when
retrieving lists of objects from the Google Drive API, this commit
refactors the `list*` methods in `google_drive.app` to be async
generators that will only query new pages on demand. This will allow
for event sources to process these changes and cache intermediate
results in order to avoid losing progress in the event of a timeout.

This fix involved:

* Refactoring the `common-webhooks` module to use the new generators
  and to own all reusable logic around fetching changes from GDrive
* Refactoring the modules that extend `common-webhooks`, which
  involved removing redundant/duplicated code, as well as changing
  the moment at which checkpoint data is cached
* Some legacy bugs that appeared during stress tests