Fix **Publish Components** GitHub action context (#1819)
Reverts changes to **Publish Components** workflow in #1810, which had changed the workflow trigger from "on push to master" to "on closed pull request".

The change in #1810 was made to enable the workflow to run on "squash/rebase and merge". But when triggered by the `pull_request` event, the workflow would run from the context of the forked repo and lose access to secrets used in the workflow.

Triggering the workflow on `push` events to the `master` branch should allow it to run in the context of the base repo on squash/rebase and merges to `master`.