Fix missing emails when mailbox is changed in IMAP New Email source
Emails after the *last seen UID* were being fetched, but email UIDs are
unique to a mailbox, not globally unique. In other words, a different
mailbox can have a more recent email with a smaller UID.

To avoid missing emails, `lastUid` and `uidValidity` are now
individually stored for each mailbox.