Change ESLint pipedream/props-label & description rules to "warn" (#2341)
Updates `pipedream/props-label` & `pipedream/props-description` ESLint
rules from "error" to "warn" to allow for commits that break these rules

Context:

Commit bb002de added a pre-commit hook which aborts commits
with changes that fail the lint check. So new or changed props without
labels or descriptions cannot be commited.

Since prop labels and descriptions aren't required by the guidelines or
for a component to be published and run successfully, they can be safely
omitted.