fix for  [warning] object was filtered for serialization in Tweet code
When _addParseId() is called on a tweet, it adds a BigInt version of the ID on the end of the object and uses it to sort, which works fine.

If you subsequently do anything to the tweet object (console.log it, update it etc), the BigInt is stripped with the warning: [warning] object was filtered for serialization  - it;s too big for the parser.

Given that it's only used for this one sort and not needed anyway, I dropped the map and did the work in the sort code. Advantage is no extra additions to the object and no map needed.

Warning fixed as a result.