* Fix undefined usernames in Slack (User) prop options
Explanation:

Only the first page of users was being fetched in async options methods
for Slack props. The fetched users were used to determine the username
that corresponds to a direct message.

This commit changes the async options methods to fetch all users (used
to create a mapping from user ID to user name).

* Refactor async options for "conversation" props

* Remove unused code from async options for the `publicChannel` prop

* Fix linter warnings in slack.app.mjs

Adds labels and descriptions to prop definitions

* Reuse conversation prop definition in Slack New Message in Channels

* Fix linter errors in Slack New Message in Channels source