Fix some users in Slack User & Channel props listed as "Undefined" in async options (#2458)
* Fix conversation prop in Slack Send Group Message action

The `conversation` prop was showing direct messages as options rather
than groups. This commit changes the conversation prop to inherit from
the `group` propDefinition.

* * Fix undefined usernames in Slack (User) prop options

Explanation:

Only the first page of users was being fetched in async options methods
for Slack props. The fetched users were used to determine the username
that corresponds to a direct message.

This commit changes the async options methods to fetch all users (used
to create a mapping from user ID to user name).

* Refactor async options for "conversation" props

* Remove unused code from async options for the `publicChannel` prop

* Fix linter warnings in slack.app.mjs

Adds labels and descriptions to prop definitions

* Reuse conversation prop definition in Slack New Message in Channels

* Bump versions of Slack components

* Add JSDocs to slack.app.mjs

* Fix - revert change removing limit param from conversations API call

* Rename Slack components with underscore to dash

Replaces "_" with "-" in component folder/file names to adhere to
component guidelines

* add_star -> add-star
* delete_file -> delete-file
* list_reminders -> list-reminders
* remove_star -> remove-star

* Fix Slack Add Star, Remove Star, Kick User actions

Renames conversation argument to channel