updated the descriptons
updated from "Emit new event when a value in the specified column is updated on a board in Monday. For changes to Name, use the Name Updated action." to "Emit new event when a value in the specified column is updated on a board in Monday. For changes to Name, use the Name Updated Trigger."