Fix GSheets New Row Added duplicate rows
* Skip emitting duplicate event for the last row in the sheet
* Fix setting row count when the new row is the last row in the sheet

Since A1 notation is inclusive of the ending row in the range, the last
row in the sheet is always included if it contains a cell with a value.
So when the event source was triggered by an update to the sheet, the
last row was always emitted as a new row when it contained a value.

For example:

If the length of Sheet1 is 25, and cell A25 contains a value, the range
to search for new rows would be Sheet1!26:25. The Google Sheets API call
for the values in the specified range would return a single row (row
25), which would be incorrectly emitted as a new row.