Using branch-specific concurrency for publish-packages workflow.
This should be ok, as github should only kill jobs that are pending if
the branch matches.  The latest commit to each branch should still run,
and must succeed for that build to succeed.

A fresh `pnpm install` and `pnpm publish` when no other publish calls are running should pick up any previously unpublished packages and publish them.