Ana06/get-changed-files v2.3.0 Considers renamed modified files as modified (#13043)
* Ana06/get-changed-files v2.3.0 Adds added_modified_renamed that includes renamed non-modified files and all files in added_modified
this caused any renamed and modified files to get added to the list of
files to publish twice, which results in a publishing error on the
second try.

* slight modification in that we should ignore renamed files that are not
modified. If a file is in fact renamed and not modified, it should cause
a different error during PR checks.

Also fixing for the TS build.