Configure file stash sync: Adobe PDF Services-Guru
Adds dir props to relevant components for apps starting with 'A' to 'G'.

The 'dir' prop indicates that a stash ID should be passed when running
these actions via the Connect API. This allows files written to the
ephemeral /tmp directory in one action to be read by another, separately
executed action since both can sync /tmp with the same File Stash
directory.