Configure file stash sync for Zoho CRM components (#17267)
Add dir props to the Zoho CRM Download Attachment and Upload Attachment
actions to indicate that a stash ID should be passed when running these
actions via the Connect API. This allows files written to the ephemeral
/tmp directory in one action to be read by another, separately executed
action since both can sync /tmp with the same File Stash directory.

The prop is optional in the Zoho CRM Upload Attachment action since the
file reference used as input can be a file URL rather than local file
path. If a file URL is used, then the /tmp <-> File Stash sync is not
needed and a stash ID is not required.

The prop is also optional in the Zoho CRM Download Attachment action
since the file content is returned directly from the `run` method of the
action, in addition to being written to a file. So the action can be
used in Connect without persisting the file for later use.