Configure file stash sync: Adobe PDF Services-Guru (#17361)
* Add dir props to relevant components for apps starting with 'A' to 'G'

The 'dir' prop indicates that a stash ID should be passed when running
these actions via the Connect API. This allows files written to the
ephemeral /tmp directory in one action to be read by another, separately
executed action since both can sync /tmp with the same File Stash
directory.

* Add loader for component app prop check

When a component imported a package with pinned version, the "Check
component app prop" PR check failed because Node.js was unable to
resolve the package version correctly. This commit adds a custom loader
that strips the version from the imported package, allowing the check to
pass.

For example, if the component imports `got@13.0.0`, the loader will
transform the import to `got`.