Upgrade to the v2 SDK (#18520)
Use the new v2 version of the Pipedream SDK in the `connect-react`
package. This change involves migrating to the new types mostly, but
also runtime changes involving the API calls.

The runtime behaviour should not be affected from a user's
perspective, except for consumers of the `connect-react` package
itself, since some components (e.g. `FrontendClientProvider`) expect
their consumers to inject a client instance of the same SDK
version. For this reason, this change bumps the **major** version of
this package.